Update MenuItem logic and constants, remove ms1.cpp
Changed MaximumNumberOfMenuItems type from size_t to int in constants.h for consistency. Updated MenuItem constructor and display logic in Menu.cpp to use the new type and correct validation. Removed the ms1.cpp tester program.

+++ project/ms1/constants.h
@@ -28,7 +28,7 @@ and include the citation of that code.
 // <<<--- END of citation CSTDDEF__CITE
 
 namespace seneca {
-   const size_t MaximumNumberOfMenuItems = 20u;
+   const int MaximumNumberOfMenuItems = 20;
 }
 
 #endif // !SENECA_CONSTANTS_H
